Clean up, add README
