+++ bsp/stm32/stm32u575-st-nucleo/.config
@@ -10,7 +10,6 @@ CONFIG_RT_NAME_MAX=8
 # CONFIG_RT_USING_ARCH_DATA_TYPE is not set
 # CONFIG_RT_USING_SMART is not set
 # CONFIG_RT_USING_AMP is not set
-# CONFIG_RT_USING_AMP is not set
 # CONFIG_RT_USING_SMP is not set
 CONFIG_RT_ALIGN_SIZE=8
 # CONFIG_RT_THREAD_PRIORITY_8 is not set
@@ -39,10 +38,6 @@ CONFIG_RT_USING_DEBUG=y
 CONFIG_RT_DEBUGING_COLOR=y
 CONFIG_RT_DEBUGING_CONTEXT=y
 CONFIG_RT_DEBUGING_INIT=y
-CONFIG_RT_USING_DEBUG=y
-CONFIG_RT_DEBUGING_COLOR=y
-CONFIG_RT_DEBUGING_CONTEXT=y
-CONFIG_RT_DEBUGING_INIT=y
 
 #
 # Inter-Thread communication
@@ -52,7 +47,6 @@ CONFIG_RT_USING_MUTEX=y
 CONFIG_RT_USING_EVENT=y
 CONFIG_RT_USING_MAILBOX=y
 CONFIG_RT_USING_MESSAGEQUEUE=y
-# CONFIG_RT_USING_MESSAGEQUEUE_PRIORITY is not set
 # CONFIG_RT_USING_MESSAGEQUEUE_PRIORITY is not set
 # CONFIG_RT_USING_SIGNALS is not set
 
@@ -240,7 +234,6 @@ CONFIG_RT_LIBC_TZ_DEFAULT_SEC=0
 # CONFIG_RT_USING_UTEST is not set
 # CONFIG_RT_USING_VAR_EXPORT is not set
 # CONFIG_RT_USING_RESOURCE_ID is not set
-# CONFIG_RT_USING_RESOURCE_ID is not set
 # CONFIG_RT_USING_ADT is not set
 # CONFIG_RT_USING_RT_LINK is not set
 # CONFIG_RT_USING_VBUS is not set
@@ -251,7 +244,7 @@ CONFIG_RT_LIBC_TZ_DEFAULT_SEC=0
 CONFIG_RT_USING_MEM_PROTECTION=y
 CONFIG_RT_USING_HW_STACK_GUARD=y
 CONFIG_NUM_MEM_REGIONS=8
-CONFIG_NUM_EXCLUSIVE_REGIONS=3
+CONFIG_NUM_EXCLUSIVE_REGIONS=2
 CONFIG_NUM_CONFIGURABLE_REGIONS=3
 
 #
