+++ libcpu/arm/cortex-m33/cpuport.c
@@ -204,7 +204,6 @@ rt_uint8_t *rt_hw_stack_init(void       *tentry,
 #else
     stack_frame->lr = 0xffffffbcL;
 #endif
-    /* TODO: Set up hardware stack guard using psplim*/
     stack_frame->psplim = 0x00;
     /*
      * CONTROL register bit assignments
@@ -240,13 +239,20 @@ rt_uint8_t *rt_hw_stack_init(void       *tentry,
 #ifdef RT_USING_HW_STACK_GUARD
 void rt_hw_stack_guard_init(rt_thread_t thread)
 {
-    rt_mem_region_t stack_top_region;
+    rt_mem_region_t stack_top_region, stack_bottom_region;
+    rt_ubase_t stack_bottom = (rt_ubase_t)thread->stack_addr;
     rt_ubase_t stack_top = (rt_ubase_t)((rt_uint8_t *)thread->stack_addr + thread->stack_size);
+    rt_ubase_t stack_bottom_region_start = RT_ALIGN(stack_bottom, MPU_MIN_REGION_SIZE);
     rt_ubase_t stack_top_region_start = RT_ALIGN_DOWN(stack_top - MPU_MIN_REGION_SIZE, MPU_MIN_REGION_SIZE);
     stack_top_region.start = (void *)stack_top_region_start;
     stack_top_region.size = MPU_MIN_REGION_SIZE;
     stack_top_region.attr = RT_MEM_REGION_P_RO_U_NA;
+    stack_bottom_region.start = (void *)stack_bottom_region_start;
+    stack_bottom_region.size = MPU_MIN_REGION_SIZE;
+    stack_bottom_region.attr = RT_MEM_REGION_P_RO_U_NA;
     rt_hw_mp_add_region(thread, &stack_top_region);
+    rt_hw_mp_add_region(thread, &stack_bottom_region);
+    thread->stack_addr = (void *)(stack_bottom_region_start + MPU_MIN_REGION_SIZE);
     thread->stack_size = (rt_uint32_t)(stack_top_region_start - (rt_ubase_t)thread->stack_addr);
 }
 #endif
