+++ libcpu/arm/cortex-m33/mpporttype.h
@@ -4,7 +4,7 @@
 #ifdef RT_USING_MEM_PROTECTION
 
 #ifdef RT_USING_HW_STACK_GUARD
-#define NUM_DYNAMIC_REGIONS (1 + NUM_CONFIGURABLE_REGIONS)
+#define NUM_DYNAMIC_REGIONS (2 + NUM_CONFIGURABLE_REGIONS)
 #else
 #define NUM_DYNAMIC_REGIONS (NUM_CONFIGURABLE_REGIONS)
 #endif
